requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
google-cloud-sql-connector
pg8000
python-dotenv